standards(render+lighting-cq): mypy --strict + ruff clean across the panda3d packages
In-file type annotations and lint fixes for the two packages allowed to import
panda3d (Hard Rule 1). No behavior change, no file/class moves. Six disjoint
file-groups fixed by a Sonnet agent fan-out (devtools_overlay; sky/weather/rain/
wind/lightning renderers; flora/grass/tree/mote/building renderers; app/registry/
gameobject/camera/transform/primitives/scene_visuals; bridges + post + profiler;
lighting/*).

With this, mypy --strict is GREEN across the whole tree (162 files, 0 errors) and
ruff + ruff format are clean across fire_engine.

Verification:
- mypy --no-incremental: Success, no issues found in 162 source files.
- ruff check fire_engine: All checks passed. ruff format --check: 162 clean.
- Runtime (render/lighting are NOT headless-tripwire-covered): tools/screenshot.py
  --frames 150 booted the full demo, stepped 150 frames, captured a valid 390 KB
  framebuffer (pixel-art grass terrain) and exited 0 with no traceback. Only the
  pre-existing iGPU RGBA16F-float-buffer warning remains.

Notes:
- Stale `# type: ignore[import]` on panda3d imports removed (panda3d stubs now
  resolve). Remaining ignores are all specific + reasoned (ShowBase no-stubs base;
  None-fallback conditional re-exports in __init__; use_cache-guarded union-attr).
- resource_adapter.register_panda_loaders now honors its IoC param
  (resource_manager._loaders.register_loader) — clears the vulture finding too.

+++ fire_engine/render/lightning_renderer.py
@@ -46,7 +46,7 @@ import math
 from typing import Any
 
 # Panda3D imports allowed in world/ per ARCHITECTURE §3.
-from panda3d.core import (  # type: ignore[import]
+from panda3d.core import (
     BoundingBox,
     ColorBlendAttrib,
     Geom,
@@ -532,8 +532,8 @@ class LightningRendererComponent(Component):
         if cover is None or not self._cover_committed:
             return None
         ox, oy = cover.origin_m
-        col = int(math.floor((x - ox) / cover.cell_m))
-        row = int(math.floor((y - oy) / cover.cell_m))
+        col = math.floor((x - ox) / cover.cell_m)
+        row = math.floor((y - oy) / cover.cell_m)
         if 0 <= col < cover.cells and 0 <= row < cover.cells:
             z = float(cover.height[row, col])
             if z > OPEN_SKY_Z * 0.5:  # a real solid voxel (not the sentinel)
